+++ Packages/com.unity.inputsystem/InputSystem/Actions/Interactions/SlowTapInteraction.cs
@@ -47,7 +47,7 @@ namespace UnityEngine.InputSystem.Interactions
             if (context.isStarted && !context.ControlIsActuated(pressPointOrDefault))
             {
                 if (context.time - m_SlowTapStartTime >= durationOrDefault)
-                    context.PerformedAndGoBackToWaiting();
+                    context.Performed();
                 else
                     ////REVIEW: does it matter to cancel right after expiration of 'duration' or is it enough to cancel on button up like here?
                     context.Canceled();